Fix base64_Decode function with a try-block.
Apparently there are two different ways that Base64 content is transmitted. Some cases
are handled by the code at the call site already, resulting in a pure Base64 block, which can
be thrown into the shell command below. Other cases result in additional headers left in the
block, which are properly stripped by this function.

The function previously tried to always strip these additional headers, resulting in an
error when they were not present. An additional try-block now handles this case.